requirements/main.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements/main.txt requirements/main.in
#
-e git+http://github.com/andrej-shaikin/fastapi-utils.git@v0.0.4#egg=fastapi-utils
    # via -r requirements/main.in
alembic==1.8.1
    # via -r requirements/main.in
amqp==5.1.1
    # via kombu
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
asyncpg==0.26.0
    # via ormar
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via -r requirements/main.in
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
databases==0.6.1
    # via
    #   -r requirements/main.in
    #   ormar
environs==9.5.0
    # via
    #   -r requirements/main.in
    #   fastapi-utils
fastapi==0.81.0
    # via
    #   -r requirements/main.in
    #   fastapi-utils
gunicorn==20.1.0
    # via -r requirements/main.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via -r requirements/main.in
idna==3.3
    # via
    #   anyio
    #   rfc3986
kombu==5.2.4
    # via celery
mako==1.2.2
    # via alembic
markupsafe==2.1.1
    # via mako
marshmallow==3.17.1
    # via environs
ormar[postgresql]==0.11.3
    # via
    #   -r requirements/main.in
    #   fastapi-utils
packaging==21.3
    # via marshmallow
prompt-toolkit==3.0.30
    # via click-repl
psycopg2-binary==2.9.3
    # via ormar
pydantic==1.10.1
    # via
    #   fastapi
    #   ormar
pyparsing==3.0.9
    # via packaging
python-dotenv==0.21.0
    # via environs
python-environ==0.4.54
    # via -r requirements/main.in
pytz==2022.2.1
    # via celery
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via click-repl
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.40
    # via
    #   alembic
    #   databases
    #   ormar
starlette==0.19.1
    # via fastapi
typing-extensions==4.3.0
    # via pydantic
uvicorn==0.18.3
    # via -r requirements/main.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
